Module Tar_eio

I/O for tar-formatted data

val get_next_header : ?level:Tar.Header.compatibility -> Eio.Flow.source -> Tar.Header.t option

Returns the next header block or None if two consecutive zero-filled blocks are discovered. Assumes stream is positioned at the possible start of a header block.

raises End_of_file
if the stream unexpectedly fails.

val header_of_file : ?level:Tar.Header.compatibility -> ?getpwuid:(int64 -> string) -> ?getgrgid:(int64 -> string) -> Eio.Fs.dir Eio.Path.t -> Tar.Header.t

Return the header needed for a particular file on disk. getpwuid and getgrgid are optional functions that should take the uid and gid respectively and return the passwd and group entry names for each. These will be added to the header.

module Archive : sig ... end

Utility functions for operating over whole tar archives